requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
filelock==3.16.1
Flask==3.1.0
grpcio==1.68.0
grpcio-tools==1.68.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.24.3
pillow==11.0.0
protobuf==5.28.3
PyYAML==6.0
requests==2.32.3
sympy==1.13.3
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
